feat: 페이지 컴포넌트를 서버 컴포넌트로 변경 후 Data Fetching
