+++ src/app/(user)/myhistory/_components/MyHistoryPageContent/MyHistorySection/HistoryCard.tsx
