--- src/app/(user)/myhistory/_components/MyHistorySection/HistoryCard/index.tsx
@@ -1,2 +1,0 @@
-import HistoryCard from './HistoryCard';
-export default HistoryCard;
